pyproject.toml
[tool.poetry]
name = "easyaptos"
version = "1.0.0"
description = "Script for various tasks in the Aptos network"
authors = ["Dellenoam"]
license = "MIT License"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
aptos-sdk = "^0.10.0"
pydantic-settings = "^2.7.0"
openpyxl = "^3.1.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
